Remove lancie-app
This is the last step required to full load the homepage on index.html
Most of the styling could be removed and only very little bindings were
used. These are replaced with simple eventListeners.

I also removed unnecessary iron-image, removed styling and other
custom elements that could simply be replaced with their
native equivalent.

We should probably remove quite some of our routes, but that
is for a future PR